requirements.txt
streamlit==1.38.0
transformers==4.24.0
pillow==10.4.0
numpy==2.1.0